Transfer API specific functions into separate packages
Also change functions so that they are specific to each API and remove GoMiler struct
